net: wireless: bcmdhd: fix ASSERT macro causing memory corruption
BUG=31089587

Since ASSERT macro translated to simple print, there's case found
where code executed after that ASSERT macro. This violation may
cause some memory corruption issue.

So removed the macro or added error return statement to prevent
further code execution.

+++ drivers/net/wireless/bcmdhd/dhd_flowring.h
@@ -172,5 +172,5 @@ extern int dhd_update_interface_link_status(dhd_pub_t *dhdp, uint8 ifindex,
 extern int dhd_flow_prio_map(dhd_pub_t *dhd, uint8 *map, bool set);
 extern int dhd_update_flow_prio_map(dhd_pub_t *dhdp, uint8 map);
 
-extern uint8 dhd_flow_rings_ifindex2role(dhd_pub_t *dhdp, uint8 ifindex);
+extern int dhd_flow_rings_ifindex2role(dhd_pub_t *dhdp, uint8 ifindex);
 #endif /* _dhd_flowrings_h_ */
